requirements.txt
pyTelegramBotAPI==3.8.3
beautifulsoup4==4.9.3
requests==2.23.0
python-dotenv==0.19.0
lxml==4.6.2
psycopg2-binary==2.9.1
schedule==1.1.0
